mysite/requirements.txt

asgiref==3.7.2
Django==5.0.9
sqlparse==0.4.4
typing_extensions==4.9.0
python-telegram-bot==20.7
Pillow==10.1.0
django-environ==0.11.2
drf-yasg==1.21.7
djangorestframework==3.14.0